requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
cairosvg==2.7.1

